refactor(ksc): split static ledger-decider + auto-loop fingerprint helpers into ParsekKSC.Decisions.cs
